VDX-181 Implement Issuer State persistence
   - Refactored interfaces;
